docs(man): add man pages for tv
